requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
httpx==0.26.0
loguru==0.7.2
aiosmtpd==1.4.4
